requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
requests>=2.28.0
numpy>=1.24.0
